Create admin interfaz

+++ salecat.sql
@@ -1,8 +1,8 @@
--- MariaDB dump 10.19  Distrib 10.9.3-MariaDB, for Linux (x86_64)
+-- MySQL dump 10.19  Distrib 10.3.34-MariaDB, for debian-linux-gnu (x86_64)
 --
 -- Host: localhost    Database: salecat
 -- ------------------------------------------------------
--- Server version	10.9.3-MariaDB
+-- Server version	10.3.34-MariaDB-0ubuntu0.20.04.1
 
 /*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
 /*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
@@ -141,7 +141,7 @@ DROP TABLE IF EXISTS `sale`;
 CREATE TABLE `sale` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `date` date NOT NULL DEFAULT curdate(),
-  `total` int(11) NOT NULL,
+  `total` float NOT NULL,
   `userFK` int(11) DEFAULT NULL,
   PRIMARY KEY (`id`),
   KEY `userFK` (`userFK`),
@@ -191,4 +191,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2022-09-26 22:48:54
+-- Dump completed on 2022-10-15 20:02:17
